logstor: extract segment reading functions to free functions
Extract the segment_manager functions that do reading and
deserialization of segments, segment headers, segment records, etc, into
free functions in a separate file that operate on a input_stream.

This improves the organization of the code and will allow us to unit
test these functions and reuse them, for tools for example.
