+++ configure.py
@@ -905,6 +905,7 @@ scylla_core = (['message/messaging_service.cc',
                 'replica/multishard_query.cc',
                 'replica/mutation_dump.cc',
                 'replica/querier.cc',
+                'replica/logstor/segment_io.cc',
                 'replica/logstor/segment_manager.cc',
                 'replica/logstor/logstor.cc',
                 'replica/logstor/write_buffer.cc',
