+++ replica/logstor/segment_io.cc
@@ -1,0 +1,214 @@
+/*
+ * Copyright (C) 2026-present ScyllaDB
+ */
+
+/*
+ * SPDX-License-Identifier: LicenseRef-ScyllaDB-Source-Available-1.1
+ */
+
+#include "replica/logstor/segment_io.hh"
+
+#include <seastar/core/align.hh>
+#include <seastar/core/simple-stream.hh>
+
+#include "idl/logstor.dist.hh"
+#include "idl/logstor.dist.impl.hh"
+#include "serializer_impl.hh"
+
+namespace replica::logstor {
+
+extern seastar::logger logstor_logger;
+
+static constexpr size_t block_alignment = 4096;
+
+segment_header make_segment_header(const write_buffer::buffer_header& bh, std::optional<write_buffer::segment_header> sh) {
+    segment_header seg_hdr {
+        .kind = bh.kind,
+        .segment_seq = bh.segment_seq,
+    };
+
+    switch (bh.kind) {
+    case segment_kind::full:
+        seg_hdr.v = segment_header::full {
+            .table = sh->table,
+            .first_token = sh->first_token,
+            .last_token = sh->last_token,
+        };
+        break;
+    case segment_kind::mixed:
+        seg_hdr.v = segment_header::mixed{};
+        break;
+    }
+    return seg_hdr;
+}
+
+future<std::optional<segment_header>> read_segment_header(seastar::input_stream<char>& in) {
+    auto bh_buf = co_await in.read_exactly(write_buffer::buffer_header_size);
+    if (bh_buf.size() < write_buffer::buffer_header_size) {
+        co_return std::nullopt;
+    }
+    auto bh = ser::deserialize_from_buffer(bh_buf, std::type_identity<write_buffer::buffer_header>{});
+
+    if (!write_buffer::validate_header(bh)) {
+        co_return std::nullopt;
+    }
+
+    std::optional<write_buffer::segment_header> sh;
+    if (bh.kind == segment_kind::full) {
+        auto sh_buf = co_await in.read_exactly(write_buffer::segment_header_size);
+        if (sh_buf.size() < write_buffer::segment_header_size) {
+            co_return std::nullopt;
+        }
+        sh = ser::deserialize_from_buffer(sh_buf, std::type_identity<write_buffer::segment_header>{});
+    }
+
+    co_return make_segment_header(bh, sh);
+}
+
+log_record deserialize_log_record(simple_memory_input_stream buf_stream) {
+    auto rh_stream = buf_stream.read_substream(write_buffer::record_header_size);
+    auto rh = ser::deserialize(rh_stream, std::type_identity<write_buffer::record_header>{});
+    auto header_stream = buf_stream.read_substream(rh.header_size);
+    auto data_stream = buf_stream.read_substream(rh.data_size);
+
+    return log_record {
+        .header = ser::deserialize(header_stream, std::type_identity<log_record_header>{}),
+        .mut = ser::deserialize(data_stream, std::type_identity<canonical_mutation>{})
+    };
+}
+
+future<log_record> read_log_record(seastar::input_stream<char>& in, log_location loc) {
+    auto buf = co_await in.read_exactly(loc.size);
+    if (buf.size() < loc.size) {
+        throw std::runtime_error(fmt::format("Truncated log record at {}", loc));
+    }
+    co_return deserialize_log_record(simple_memory_input_stream(buf.begin(), buf.size()));
+}
+
+future<> scan_segment(seastar::input_stream<char>& in,
+        log_segment_id segment_id,
+        size_t segment_size,
+        segment_header_consumer on_segment_header,
+        record_header_consumer on_record_header,
+        record_consumer on_record) {
+    size_t current_position = 0;
+    std::optional<segment_sequence> segment_seq;
+
+    logstor_logger.trace("Reading records from segment {}", segment_id);
+
+    while (current_position < segment_size) {
+        // Align to block boundary
+        auto skip_bytes = align_up(current_position, block_alignment) - current_position;
+        if (skip_bytes > 0) {
+            co_await in.skip(skip_bytes);
+            current_position += skip_bytes;
+        }
+
+        if (current_position >= segment_size) {
+            break;
+        }
+
+        // read buffer header
+        auto buffer_header_buf = co_await in.read_exactly(write_buffer::buffer_header_size);
+        current_position += write_buffer::buffer_header_size;
+        if (buffer_header_buf.size() < write_buffer::buffer_header_size) {
+            break;
+        }
+        auto bh = ser::deserialize_from_buffer(buffer_header_buf, std::type_identity<write_buffer::buffer_header>{});
+
+        // if the buffer is invalid then skip the rest of the segment - buffer writes are sequential and serialized.
+        if (!write_buffer::validate_header(bh)) {
+            break;
+        }
+
+        if (!segment_seq) {
+            segment_seq = bh.segment_seq;
+        } else if (bh.segment_seq != *segment_seq) {
+            break;
+        }
+
+        std::optional<write_buffer::segment_header> sh;
+        if (bh.kind == segment_kind::full) {
+            // read segment header
+            auto segment_header_buf = co_await in.read_exactly(write_buffer::segment_header_size);
+            current_position += write_buffer::segment_header_size;
+            if (segment_header_buf.size() < write_buffer::segment_header_size) {
+                break;
+            }
+            sh = ser::deserialize_from_buffer(segment_header_buf, std::type_identity<write_buffer::segment_header>{});
+        }
+
+        auto seg_hdr = make_segment_header(bh, sh);
+        co_await on_segment_header(seg_hdr);
+
+        // TODO crc, torn writes
+
+        const auto buffer_data_end_position = current_position + bh.data_size;
+        while (current_position < buffer_data_end_position) {
+            // Read record header
+            const auto record_offset = current_position;
+            auto size_buf = co_await in.read_exactly(write_buffer::record_header_size);
+            current_position += write_buffer::record_header_size;
+            if (size_buf.size() < write_buffer::record_header_size) {
+                break;
+            }
+            auto rh = ser::deserialize_from_buffer(size_buf, std::type_identity<write_buffer::record_header>{});
+            if (rh.header_size == 0 || rh.header_size + rh.data_size > buffer_data_end_position - current_position) {
+                // invalid record size
+                break;
+            }
+
+            logstor_logger.trace("Found record of size {} bytes in segment {}",
+                                rh.header_size + rh.data_size, segment_id);
+
+            // Read the log_record_header bytes
+            auto header_buf = co_await in.read_exactly(rh.header_size);
+            current_position += rh.header_size;
+            if (header_buf.size() < rh.header_size) {
+                break;
+            }
+            auto record_header = ser::deserialize_from_buffer(header_buf, std::type_identity<log_record_header>{});
+
+            log_location loc {
+                .segment = segment_id,
+                .offset = static_cast<uint32_t>(record_offset),
+                .size = static_cast<uint32_t>(write_buffer::record_header_size + rh.header_size + rh.data_size)
+            };
+
+            if (on_record_header(loc, record_header) == want_data::yes) {
+                auto mut_buf = co_await in.read_exactly(rh.data_size);
+                current_position += rh.data_size;
+                if (mut_buf.size() < rh.data_size) {
+                    break;
+                }
+                auto mut = ser::deserialize_from_buffer(mut_buf, std::type_identity<canonical_mutation>{});
+                co_await on_record(loc, log_record{std::move(record_header), std::move(mut)});
+            } else {
+                // Skip the canonical_mutation bytes without reading them
+                co_await in.skip(rh.data_size);
+                current_position += rh.data_size;
+            }
+
+            // align up to next record
+            auto padding = align_up(current_position, write_buffer::record_alignment) - current_position;
+            if (padding > 0) {
+                co_await in.skip(padding);
+                current_position += padding;
+            }
+        }
+
+        if (seg_hdr.kind == segment_kind::full) {
+            // A segment of this kind has only a single buffer
+            break;
+        }
+
+        if (current_position < buffer_data_end_position) {
+            // skip remaining buffer data
+            auto bytes_to_skip = buffer_data_end_position - current_position;
+            co_await in.skip(bytes_to_skip);
+            current_position += bytes_to_skip;
+        }
+    }
+}
+
+}
