+++ replica/logstor/segment_manager.hh
@@ -19,6 +19,7 @@
 #include <seastar/util/bool_class.hh>
 #include "bytes_fwd.hh"
 #include "mutation_writer/token_group_based_splitting_writer.hh"
+#include "replica/logstor/segment_io.hh"
 #include "replica/logstor/write_buffer.hh"
 #include "replica/logstor/compaction.hh"
 #include "types.hh"
@@ -30,8 +31,6 @@ class database;
 class table;
 
 namespace logstor {
-
-using want_data = seastar::bool_class<class want_data_tag>;
 
 class compaction_manager;
 class segment_set;
